+++ test/cim/iec61970/base/wires/test_conductor.py
@@ -7,8 +7,9 @@ from hypothesis.strategies import floats
 
 from test.cim.iec61970.base.core.test_conducting_equipment import conducting_equipment_kwargs, verify_conducting_equipment_constructor_default, \
     verify_conducting_equipment_constructor_kwargs, verify_conducting_equipment_constructor_args, conducting_equipment_args
+from test.cim.property_validator import validate_property_accessor
 from test.cim_creators import FLOAT_MIN, FLOAT_MAX
-from zepben.evolve import Conductor
+from zepben.evolve import Conductor, WireInfo
 
 conductor_kwargs = {
     **conducting_equipment_kwargs,
@@ -31,3 +32,7 @@ def verify_conductor_constructor_kwargs(c: Conductor, length, **kwargs):
 def verify_conductor_constructor_args(c: Conductor):
     verify_conducting_equipment_constructor_args(c)
     assert c.length == conductor_args[-1]
+
+
+def test_wire_info_accessor():
+    validate_property_accessor(Conductor, WireInfo, Conductor.wire_info)