+++ test/cim_creators.py
@@ -91,6 +91,17 @@ def create_short_circuit_test():
     )
 
 
+def create_shunt_compensator_info():
+    return builds(
+        ShuntCompensatorInfo,
+        **create_asset_info(),
+        max_power_loss=integers(min_value=MIN_32_BIT_INTEGER, max_value=MAX_32_BIT_INTEGER),
+        rated_current=integers(min_value=MIN_32_BIT_INTEGER, max_value=MAX_32_BIT_INTEGER),
+        rated_reactive_power=integers(min_value=MIN_32_BIT_INTEGER, max_value=MAX_32_BIT_INTEGER),
+        rated_voltage=integers(min_value=MIN_32_BIT_INTEGER, max_value=MAX_32_BIT_INTEGER),
+    )
+
+
 def create_transformer_end_info():
     return builds(
         TransformerEndInfo,
@@ -869,6 +880,7 @@ def create_regulating_cond_eq():
 def create_shunt_compensator():
     return {
         **create_regulating_cond_eq(),
+        "asset_info": builds(ShuntCompensatorInfo, **create_identified_object()),
         "sections": floats(min_value=FLOAT_MIN, max_value=FLOAT_MAX),
         "grounded": booleans(),
         "nom_u": integers(min_value=MIN_32_BIT_INTEGER, max_value=MAX_32_BIT_INTEGER),