+++ test/services/network/test_network_service_comparator.py
@@ -6,8 +6,6 @@
 
 from typing import Type
 
-from test.services.common.service_comparator_validator import ServiceComparatorValidator
-from test.services.common.test_base_service_comparator import TestBaseServiceComparator
 from zepben.evolve import CableInfo, NoLoadTest, OpenCircuitTest, OverheadWireInfo, PowerTransformerInfo, TransformerTankInfo, ShortCircuitTest, \
     TransformerEndInfo, TransformerStarImpedance, TransformerTest, WireInfo, WireMaterialKind, Asset, AssetOwner, Location, AssetContainer, AssetInfo, \
     AssetOrganisationRole, Pole, Streetlight, WindingConnection, StreetlightLampKind, Structure, StreetAddress, TownDetail, PositionPoint, EndDevice, \
@@ -19,8 +17,11 @@ from zepben.evolve import CableInfo, NoLoadTest, OpenCircuitTest, OverheadWireIn
     EnergyConsumer, PhaseShuntConnectionKind, EnergyConsumerPhase, SinglePhaseKind, EnergySource, EnergySourcePhase, Fuse, Jumper, Line, \
     LinearShuntCompensator, PerLengthImpedance, PerLengthLineParameter, PowerElectronicsConnectionPhase, PowerTransformer, PowerTransformerEnd, VectorGroup, \
     ProtectedSwitch, RatioTapChanger, Recloser, RegulatingCondEq, ShuntCompensator, Switch, ObjectDifference, ValueDifference, TapChanger, TransformerEnd, \
-    Circuit, Loop, NetworkService, TracedPhases, PhaseDirection
+    Circuit, Loop, NetworkService, TracedPhases, PhaseDirection, ShuntCompensatorInfo
 from zepben.evolve.services.network.network_service_comparator import NetworkServiceComparatorOptions, NetworkServiceComparator
+
+from test.services.common.service_comparator_validator import ServiceComparatorValidator
+from test.services.common.test_base_service_comparator import TestBaseServiceComparator
 
 
 class TestNetworkServiceComparator(TestBaseServiceComparator):
@@ -78,6 +79,13 @@ class TestNetworkServiceComparator(TestBaseServiceComparator):
         self.validator.validate_property(ShortCircuitTest.power, ShortCircuitTest, lambda _: 1, lambda _: 2)
         self.validator.validate_property(ShortCircuitTest.voltage, ShortCircuitTest, lambda _: 1.0, lambda _: 2.0)
         self.validator.validate_property(ShortCircuitTest.voltage_ohmic_part, ShortCircuitTest, lambda _: 1.0, lambda _: 2.0)
+
+    def test_compare_shunt_compensator_info(self):
+        self._compare_asset_info(ShuntCompensatorInfo)
+        self.validator.validate_property(ShuntCompensatorInfo.max_power_loss, ShuntCompensatorInfo, lambda _: 1, lambda _: 2)
+        self.validator.validate_property(ShuntCompensatorInfo.rated_current, ShuntCompensatorInfo, lambda _: 1, lambda _: 2)
+        self.validator.validate_property(ShuntCompensatorInfo.rated_reactive_power, ShuntCompensatorInfo, lambda _: 1, lambda _: 2)
+        self.validator.validate_property(ShuntCompensatorInfo.rated_voltage, ShuntCompensatorInfo, lambda _: 1, lambda _: 2)
 
     def test_compare_transformer_end_info(self):
         self._compare_asset_info(TransformerEndInfo)
@@ -816,6 +824,8 @@ class TestNetworkServiceComparator(TestBaseServiceComparator):
             lambda _: PhaseShuntConnectionKind.G
         )
         self.validator.validate_property(ShuntCompensator.sections, creator, lambda _: 1.0, lambda _: 2.0)
+        self.validator.validate_property(ShuntCompensator.asset_info, creator, lambda _: ShuntCompensatorInfo(mrid="sci1"),
+                                         lambda _: ShuntCompensatorInfo(mrid="sci2"))
 
     def _compare_switch(self, creator: Type[Switch]):
         self._compare_conducting_equipment(creator)